pyproject.toml
[project]
name = "ads_mod_11_qualidade_e_testes_de_software"
version = "1.0.0"
description = "Qualidade e Testes de Software"
authors = [
    {name = "Ricardo Tec Pro", email = "ricardotecpro@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.6.21,<10.0.0)",
    "mkdocs-glightbox (>=0.5.1,<0.6.0)",
    "mkdocs-quiz (>=0.0.41,<0.0.42)",
    "mkdocs-macros-plugin (>=1.4.0,<2.0.0)",
    "mkdocs-git-authors-plugin (>=0.10.0,<0.11.0)",
    "mkdocs-git-revision-date-localized-plugin (>=1.4.7,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pymdown-extensions (>=10.0.0,<11.0.0)",
    "mkdocs-video (>=1.5.0,<2.0.0)",
    "termynal (>=0.13.1,<0.14.0)",
    "mkdocs-minify-plugin (>=0.8.0,<0.9.0)",
    "mkdocs-print-site-plugin (>=2.8,<3.0)",
    "mkdocs-awesome-pages-plugin (>=2.10.1,<3.0.0)",
    "mkdocs-jupyter (>=0.25.1,<0.26.0)",
    "mkdocs-rss-plugin (>=1.17.9,<2.0.0)",
    "mkdocs-revealjs (>=0.0.6,<0.0.7)",
    "PyYAML (>=6.0.1,<7.0.0)",
]

[dependency-groups]
dev = [
  "invoke (>=2.2.0,<3.0.0)",
  "rich (>=14.1.0,<15.0.0)",
  "taskipy (>=1.14.1,<2.0.0)",
  "pytest (>=8.0.0,<9.0.0)",
  "pytest-playwright (>=0.5.0,<0.6.0)",
  "playwright (>=1.40.0,<2.0.0)",
  "ruff (>=0.15.0,<0.16.0)",
]

[tool.taskipy.tasks]
slides = "python scripts/generate_slides_quizzes.py"
quizzes = "python scripts/convert_quizzes.py"
exercises = "python scripts/build_exercises.py"
build = "mkdocs build --strict"
audit = "python d:/SourceCode/REPOS/_prompts/scripts/unified_audit.py ."
test = "pytest tests/ -v"
serve = "mkdocs serve"
deploy = "mkdocs gh-deploy --force"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
base_url = "http://127.0.0.1:8766"
